requirements.txt
Flask
torch
torchvision
pandas
Pillow